MongoIterables simplification
No longer need runtime specialization of iterables now that Java 8 is
the minimum

JAVA-2996
