+++ driver-sync/src/test/unit/com/mongodb/client/internal/MongoIterablesSpecification.groovy
@@ -17,17 +17,20 @@
 package com.mongodb.client.internal
 
 import com.mongodb.MongoClientSettings
+import com.mongodb.MongoInternalException
 import com.mongodb.MongoNamespace
 import com.mongodb.ReadConcern
 import com.mongodb.WriteConcern
 import com.mongodb.client.ClientSession
 import com.mongodb.client.model.AggregationLevel
 import com.mongodb.client.model.changestream.ChangeStreamLevel
+import com.mongodb.operation.BatchCursor
 import org.bson.BsonBoolean
 import org.bson.BsonDocument
 import org.bson.Document
-import spock.lang.IgnoreIf
 import spock.lang.Specification
+
+import java.util.function.Consumer
 
 import static com.mongodb.CustomMatchers.isTheSameAs
 import static com.mongodb.ReadPreference.secondary
@@ -35,101 +38,18 @@ import static spock.util.matcher.HamcrestSupport.expect
 
 class MongoIterablesSpecification extends Specification {
 
-    def executor = new TestOperationExecutor([])
+    static namespace = new MongoNamespace('databaseName', 'collectionName')
+    static codecRegistry = MongoClientSettings.getDefaultCodecRegistry()
+    static readPreference = secondary()
+    static readConcern = ReadConcern.MAJORITY
+    static writeConcern = WriteConcern.MAJORITY
+    static filter = new BsonDocument('x', BsonBoolean.TRUE)
+    static pipeline = Collections.emptyList()
+
     def clientSession = Stub(ClientSession)
-    def namespace = new MongoNamespace('databaseName', 'collectionName')
-    def codecRegistry = MongoClientSettings.getDefaultCodecRegistry()
-    def readPreference = secondary()
-    def readConcern = ReadConcern.MAJORITY
-    def writeConcern = WriteConcern.MAJORITY
-    def filter = new BsonDocument('x', BsonBoolean.TRUE)
-    def pipeline = Collections.emptyList()
+    def executor = new TestOperationExecutor([])
 
-    @IgnoreIf({ !Java8MongoIterablesSpecification.IS_CONSUMER_CLASS_AVAILABLE })
-    def 'should create Java 8 iterables when java.util.function.Consumer is available'() {
-        when:
-        def findIterable = MongoIterables.findOf(clientSession, namespace, Document, BsonDocument, codecRegistry, readPreference,
-                readConcern, executor, filter)
-
-        then:
-        expect findIterable, isTheSameAs(new Java8FindIterableImpl<Document, BsonDocument>(clientSession, namespace, Document,
-                BsonDocument, codecRegistry, readPreference, readConcern, executor, filter))
-
-        when:
-        def aggregateIterable = MongoIterables.aggregateOf(clientSession, namespace, Document, BsonDocument, codecRegistry,
-                readPreference, readConcern, writeConcern, executor, pipeline, AggregationLevel.COLLECTION)
-
-        then:
-        expect aggregateIterable, isTheSameAs(new Java8AggregateIterableImpl<Document, BsonDocument>(clientSession, namespace,
-                Document, BsonDocument, codecRegistry, readPreference, readConcern, writeConcern, executor, pipeline,
-                AggregationLevel.COLLECTION))
-
-        when:
-        aggregateIterable = MongoIterables.aggregateOf(clientSession, namespace.databaseName, Document, BsonDocument, codecRegistry,
-                readPreference, readConcern, writeConcern, executor, pipeline, AggregationLevel.DATABASE)
-
-        then:
-        expect aggregateIterable, isTheSameAs(new Java8AggregateIterableImpl<Document, BsonDocument>(clientSession, namespace.databaseName,
-                Document, BsonDocument, codecRegistry, readPreference, readConcern, writeConcern, executor, pipeline,
-                AggregationLevel.DATABASE))
-
-        when:
-        def changeStreamIterable = MongoIterables.changeStreamOf(clientSession, namespace, codecRegistry,
-                readPreference, readConcern, executor, pipeline, Document, ChangeStreamLevel.COLLECTION)
-
-        then:
-        expect changeStreamIterable, isTheSameAs(new Java8ChangeStreamIterableImpl(clientSession, namespace, codecRegistry,
-                readPreference, readConcern, executor, pipeline, Document, ChangeStreamLevel.COLLECTION), ['codec'])
-
-        when:
-        changeStreamIterable = MongoIterables.changeStreamOf(clientSession, namespace.databaseName, codecRegistry,
-                readPreference, readConcern, executor, pipeline, Document, ChangeStreamLevel.COLLECTION)
-
-        then:
-        expect changeStreamIterable, isTheSameAs(new Java8ChangeStreamIterableImpl(clientSession, namespace.databaseName, codecRegistry,
-                readPreference, readConcern, executor, pipeline, Document, ChangeStreamLevel.COLLECTION), ['codec'])
-
-        when:
-        def distinctIterable = MongoIterables.distinctOf(clientSession, namespace, Document, BsonDocument, codecRegistry, readPreference,
-                readConcern, executor, 'f1', filter)
-
-        then:
-        expect distinctIterable, isTheSameAs(new Java8DistinctIterableImpl(clientSession, namespace, Document, BsonDocument, codecRegistry,
-                readPreference, readConcern, executor, 'f1', filter))
-
-        when:
-        def listDatabasesIterable = MongoIterables.listDatabasesOf(clientSession, Document, codecRegistry, readPreference, executor)
-
-        then:
-        expect listDatabasesIterable, isTheSameAs(new Java8ListDatabasesIterableImpl(clientSession, Document, codecRegistry, readPreference,
-                executor))
-
-        when:
-        def listCollectionsIterable = MongoIterables.listCollectionsOf(clientSession, 'test', true, Document,
-                codecRegistry, readPreference, executor)
-
-        then:
-        expect listCollectionsIterable, isTheSameAs(new Java8ListCollectionsIterableImpl(clientSession, 'test', true, Document,
-                codecRegistry, readPreference, executor))
-
-        when:
-        def listIndexesIterable = MongoIterables.listIndexesOf(clientSession, namespace, Document, codecRegistry, readPreference, executor)
-
-        then:
-        expect listIndexesIterable, isTheSameAs(new Java8ListIndexesIterableImpl(clientSession, namespace, Document, codecRegistry,
-                readPreference, executor))
-
-        when:
-        def mapReduceIterable = MongoIterables.mapReduceOf(clientSession, namespace, Document, BsonDocument, codecRegistry, readPreference,
-                readConcern, writeConcern, executor, 'map', 'reduce')
-
-        then:
-        expect mapReduceIterable, isTheSameAs(new Java8MapReduceIterableImpl(clientSession, namespace, Document, BsonDocument,
-                codecRegistry, readPreference, readConcern, writeConcern, executor, 'map', 'reduce'))
-    }
-
-    @IgnoreIf({ Java8MongoIterablesSpecification.IS_CONSUMER_CLASS_AVAILABLE })
-    def 'should create non-Java 8 iterables when java.util.function.Consumer is unavailable'() {
+    def 'should create iterables'() {
         when:
         def findIterable = MongoIterables.findOf(clientSession, namespace, Document, BsonDocument, codecRegistry, readPreference,
                 readConcern, executor, filter)
@@ -208,5 +128,102 @@ class MongoIterablesSpecification extends Specification {
 
         then:
         expect mapReduceIterable, isTheSameAs(new MapReduceIterableImpl(clientSession, namespace, Document, BsonDocument,
-                codecRegistry, readPreference, readConcern, writeConcern, executor, 'map', 'reduce'))    }
+                codecRegistry, readPreference, readConcern, writeConcern, executor, 'map', 'reduce'))
+    }
+
+    def 'should properly override Iterable forEach of Consumer'() {
+        given:
+        def cannedResults = [new Document('_id', 1), new Document('_id', 2), new Document('_id', 3)]
+        def isClosed = false
+        def count = 0
+        def accepted = 0
+
+        def cursor = {
+            Stub(BatchCursor) {
+                def results;
+                def getResult = {
+                    count++
+                    results = count == 1 ? cannedResults : null
+                    results
+                }
+                next() >> {
+                    getResult()
+                }
+                hasNext() >> {
+                    count == 0
+                }
+                close() >> {
+                    isClosed = true
+                }
+            }
+        }
+
+        when:
+        count = 0
+        accepted = 0
+        def mongoIterable = mongoIterableFactory(new TestOperationExecutor([cursor(), cursor(), cursor(), cursor()]))
+        mongoIterable.forEach(new Consumer<Document>() {
+            @Override
+            void accept(Document document) {
+                accepted++
+            }
+        })
+
+        then:
+        accepted == 3
+        isClosed
+
+        when:
+        count = 0
+        accepted = 0
+        mongoIterable = mongoIterableFactory(new TestOperationExecutor([cursor(), cursor(), cursor(), cursor()]))
+        mongoIterable.forEach(new Consumer<Document>() {
+            @Override
+            void accept(Document document) {
+                accepted++
+                throw new MongoInternalException("I don't accept this")
+            }
+        })
+
+        then:
+        thrown(MongoInternalException)
+        accepted == 1
+        isClosed
+
+        where:
+        mongoIterableFactory << [
+                { executor ->
+                    new FindIterableImpl(null, namespace, Document, Document, codecRegistry,
+                            readPreference, readConcern, executor, filter)
+                },
+                { executor ->
+                    new AggregateIterableImpl(null, namespace, Document, Document, codecRegistry,
+                            readPreference, readConcern, writeConcern, executor, pipeline, AggregationLevel.COLLECTION)
+                },
+                { executor ->
+                    new ChangeStreamIterableImpl(null, namespace, codecRegistry, readPreference, readConcern, executor, pipeline,
+                            Document, ChangeStreamLevel.COLLECTION)
+                },
+                { executor ->
+                    new DistinctIterableImpl(null, namespace, Document, Document, codecRegistry, readPreference, readConcern,
+                            executor, 'f1', filter)
+                },
+                { executor ->
+                    new ListDatabasesIterableImpl(null, Document, codecRegistry, readPreference,
+                            executor)
+                },
+                { executor ->
+                    new ListCollectionsIterableImpl(null, 'test', true, Document,
+                            codecRegistry, readPreference, executor)
+                },
+                { executor ->
+                    new ListIndexesIterableImpl(null, namespace, Document, codecRegistry,
+                            readPreference, executor)
+                },
+                { executor ->
+                    new MapReduceIterableImpl(null, namespace, Document, BsonDocument,
+                            codecRegistry, readPreference, readConcern, writeConcern, executor, 'map', 'reduce')
+                }
+        ]
+    }
 }
